server files moved to their own folder
