+++ README.md
@@ -17,11 +17,10 @@ Today Age of Merchants is a fast-growing project that became much better every d
 5. Rename all .json.default to .json in config/ folder
 6. Configure database connection in config/config.json
 7. node_modules/.bin/sequelize db:migrate
-8. $ webpack --watch
 9. (optional)  $ gen region "Region name" "X" "Y"
 ```
     Example:   $ gen region MyRegion 0 0
 ```
-10. $ node bin/www (runs the server)
+10. $ npm start
 
 That's it! I hope I didn't forget something, if I did, plz let me know in issues!
