+++ package.json
@@ -3,7 +3,7 @@
   "version": "0.0.0",
   "private": true,
   "scripts": {
-    "start": "DEBUG=aom ./bin/www",
+    "start": "webpack-dev-server --content-base public/ --port 3000",
     "test": "node tests/battle.js"
   },
   "dependencies": {
@@ -21,6 +21,7 @@
     "extract-text-webpack-plugin": "^1.0.1",
     "file-loader": "^0.9.0",
     "file-stream-rotator": "0.0.6",
+    "html-webpack-plugin": "^2.22.0",
     "jade": "~1.11.0",
     "jade-loader": "^0.8.0",
     "jsesc": "^0.5.0",
@@ -37,6 +38,8 @@
     "style-loader": "^0.13.1",
     "test": "^0.6.0",
     "typescript": "^2.0.3",
-    "webpack": "^1.13.2"
+    "webpack": "^1.13.2",
+    "webpack-dev-server": "^1.16.1",
+    "webpack-shell-plugin": "^0.4.3"
   }
 }
