+++ server/lib/auth.js
