+++ server/lib/battle.js
