+++ server/lib/game/contracts.js
