+++ server/lib/game/troops.js
