+++ server/models/soldiers.js
