+++ server/models/troops.js
