+++ server/models/troops_attacks.js
