+++ server/routes/locales.js
