+++ server/routes/registration.js
