+++ views/index.jade
@@ -23,4 +23,3 @@ html
       a.item(ng-href='#auth' ng-hide='user.id > 0') {{loc.site.signin}}
       a.item(ng-href='#auth' ng-hide='user.id > 0') {{loc.site.signup}}
   div.view(ng-view)
-  script(src='/js/app.bundle.js')
